Add RuntimeSpec
The RuntimeSpec will define the "type" of container, and what kinds of
resources it should hold. I came up with a bare list. It can easily change
later. Additionally, I made some default parameters for the cpu, memory,
and storage components of the ResourceSpec (this was renamed from
Runtime Resource)