Action uploads (#6)
Allow uploading actions that use the new IR definition. Preserves backwards compatibility to upload the previous action definitions. Action activations are not yet supported for the new IR uploads

* begin IR implementation

First, we need to define the objects based on the IR definition. To keep
with the style of the rest of the codebase the application/function
scala types are named WhiskApplication and WhiskFunction. To prevent
and unecessarily large refactor on the "WhiskAction" type - I am going
to leave that name as is, but know that it represents the "IR Object":
the base execution primitive in this system

Next steps will include modifying the REST API so that new applications
can  be uploaded uploaded to the system. After which, we will work on
the execution flow.

* tmp commit

* Update whisk definitions with new fields for IR

* Improve logging and fix serialization of actions

* Fix equality failures in ActionsApiTests

* Put whisk application into openwhisk

* Add parentFunc reference to actions

* Remove optionals from function parent